refactor(LottoPrize): enums 패키지 생성
